[Feat] 푸시알림 로직 변경 (토글 할 때)
